refactor(control-plane): extract session repository row
