+++ packages/control-plane/src/session/repository-target.test.ts
@@ -1,6 +1,6 @@
 import { RepositoryPairValidationError } from "@open-inspect/shared";
 import { describe, expect, it } from "vitest";
-import type { SessionRepositoryRow } from "./repository";
+import type { SessionRepositoryRow } from "./repository-row";
 import {
   AmbiguousRepositoryTargetError,
   buildSessionRepositories,
