+++ packages/control-plane/src/session/repository-target.ts
@@ -2,7 +2,7 @@ import {
   normalizeOptionalRepositoryPair,
   RepositoryPairValidationError,
 } from "@open-inspect/shared";
-import type { SessionRepositoryRow } from "./repository";
+import type { SessionRepositoryRow } from "./repository-row";
 
 /** A repository identified by owner and name (canonical casing unless noted). */
 export interface RepoIdentity {
